+++ test/node/node_test.cpp
@@ -48,7 +48,7 @@ struct NodeDumpParam {
 
 class NodeDumpTest : public ::testing::TestWithParam<NodeDumpParam>, public TempFileFactory {
 protected:
-  static constexpr unsigned int GVAR_NUM = 52;
+  static constexpr unsigned int GVAR_NUM = 53;
 
   NodeDumpParam param;
 
@@ -928,7 +928,7 @@ nodes:
         size: 1
       name: "f"
     handle:
-      index: 52
+      index: 53
       kind: "UDC"
       attribute: "READ_ONLY | GLOBAL"
     paramNode: null
@@ -957,7 +957,7 @@ nodes:
             type: "Int"
             varName: "?"
             handle:
-              index: 17
+              index: 18
               kind: "VAR"
               attribute: "GLOBAL"
             extraOp: "NONE"
@@ -1169,7 +1169,7 @@ nodes:
     attrNodes:
     maxVarNum: 2
     handle:
-      index: 52
+      index: 53
       kind: "FUNC"
       attribute: "READ_ONLY | GLOBAL"
     resolvedType: "(Int, Int) -> Void"
@@ -1263,7 +1263,7 @@ nodes:
                 type: "() -> Void"
                 varName: "gg"
                 handle:
-                  index: 53
+                  index: 54
                   kind: "FUNC"
                   attribute: "READ_ONLY | GLOBAL"
                 extraOp: "NONE"
@@ -1303,7 +1303,7 @@ nodes:
         attrNodes:
         maxVarNum: 0
         handle:
-          index: 52
+          index: 53
           kind: "FUNC"
           attribute: "READ_ONLY | GLOBAL"
         resolvedType: "() -> Void"
@@ -1358,7 +1358,7 @@ nodes:
         attrNodes:
         maxVarNum: 0
         handle:
-          index: 53
+          index: 54
           kind: "FUNC"
           attribute: "READ_ONLY | GLOBAL"
         resolvedType: "() -> Void"